Extract Item.Load and Item.TemplateChange
