+++ src/commands/selection.js
@@ -2,7 +2,7 @@
 
 const { call, put, select } = require('redux-saga/effects')
 const { Command } = require('./command')
-const { SaveCommand } = require('./subject')
+const subject = require('./subject')
 const { Image } = require('../image')
 const mod = require('../models')
 const act = require('../actions')
@@ -181,7 +181,7 @@ class Save extends Command {
 Save.register(SELECTION.SAVE)
 
 
-class TemplateChange extends SaveCommand {
+class TemplateChange extends subject.Save {
   type = 'selection'
 }
